# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TaskCoin API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: users
    description: User management
  - name: plans
    description: Investment plans
  - name: tasks
    description: Daily tasks
  - name: transactions
    description: Deposits and withdrawals
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /plans:
    get:
      operationId: getPlans
      tags: [plans]
      summary: Get all investment plans
      responses:
        "200":
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"

  /users/plan:
    post:
      operationId: activatePlan
      tags: [users]
      summary: Activate a plan for the current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivatePlanRequest"
      responses:
        "200":
          description: Plan activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks:
    get:
      operationId: getMyTasks
      tags: [tasks]
      summary: Get today's tasks for the current user
      responses:
        "200":
          description: Today's tasks with countdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TasksResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tasks/{taskId}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Complete a task and earn reward
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteTaskResponse"
        "400":
          description: Already completed or not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /transactions:
    get:
      operationId: getMyTransactions
      tags: [transactions]
      summary: Get transaction history for current user
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a deposit or withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get platform statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"

  /admin/users/{userId}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get a single user detail
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"

  /admin/users/{userId}/suspend:
    post:
      operationId: suspendUser
      tags: [admin]
      summary: Suspend or reactivate a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuspendUserRequest"
      responses:
        "200":
          description: Done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /admin/users/{userId}/bonus:
    post:
      operationId: addBonus
      tags: [admin]
      summary: Add bonus to user balance
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBonusRequest"
      responses:
        "200":
          description: Bonus added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /admin/transactions:
    get:
      operationId: getAdminTransactions
      tags: [admin]
      summary: Get all transactions
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminTransaction"

  /admin/transactions/{txId}/validate:
    post:
      operationId: validateTransaction
      tags: [admin]
      summary: Validate or reject a transaction
      parameters:
        - name: txId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateTransactionRequest"
      responses:
        "200":
          description: Done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get site settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    post:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update site settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettings"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        username:
          type: string
      required: [email, password, username]

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user, message]

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        balance:
          type: number
        isAdmin:
          type: boolean
        isSuspended:
          type: boolean
        activePlanId:
          type: integer
          nullable: true
        planActivatedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, email, username, balance, isAdmin, isSuspended]

    Plan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        depositRequired:
          type: number
        tasksPerDay:
          type: integer
        gainPerTask:
          type: number
        totalPerDay:
          type: number
      required: [id, name, depositRequired, tasksPerDay, gainPerTask, totalPerDay]

    ActivatePlanRequest:
      type: object
      properties:
        planId:
          type: integer
      required: [planId]

    Task:
      type: object
      properties:
        id:
          type: integer
        taskNumber:
          type: integer
        completed:
          type: boolean
        completedAt:
          type: string
          nullable: true
        gain:
          type: number
      required: [id, taskNumber, completed, gain]

    TasksResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
        resetAt:
          type: string
        secondsUntilReset:
          type: integer
        plan:
          $ref: "#/components/schemas/Plan"
          nullable: true
      required: [tasks, resetAt, secondsUntilReset]

    CompleteTaskResponse:
      type: object
      properties:
        message:
          type: string
        gain:
          type: number
        newBalance:
          type: number
      required: [message, gain, newBalance]

    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal]
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        txHash:
          type: string
          nullable: true
        walletAddress:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, type, amount, currency, status, createdAt, updatedAt]

    CreateTransactionRequest:
      type: object
      properties:
        type:
          type: string
          enum: [deposit, withdrawal]
        amount:
          type: number
        currency:
          type: string
          enum: [USDT, TRX]
        txHash:
          type: string
          nullable: true
        walletAddress:
          type: string
          nullable: true
      required: [type, amount, currency]

    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        activeUsers:
          type: integer
        pendingDeposits:
          type: integer
        pendingWithdrawals:
          type: integer
        totalDeposited:
          type: number
        totalWithdrawn:
          type: number
      required: [totalUsers, activeUsers, pendingDeposits, pendingWithdrawals, totalDeposited, totalWithdrawn]

    AdminUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        balance:
          type: number
        isAdmin:
          type: boolean
        isSuspended:
          type: boolean
        activePlanId:
          type: integer
          nullable: true
        planName:
          type: string
          nullable: true
        createdAt:
          type: string
        totalDeposited:
          type: number
        totalWithdrawn:
          type: number
      required: [id, email, username, balance, isAdmin, isSuspended, createdAt, totalDeposited, totalWithdrawn]

    SuspendUserRequest:
      type: object
      properties:
        suspended:
          type: boolean
      required: [suspended]

    AddBonusRequest:
      type: object
      properties:
        amount:
          type: number
        note:
          type: string
      required: [amount]

    AdminTransaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        email:
          type: string
        type:
          type: string
          enum: [deposit, withdrawal]
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        txHash:
          type: string
          nullable: true
        walletAddress:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, username, email, type, amount, currency, status, createdAt, updatedAt]

    ValidateTransactionRequest:
      type: object
      properties:
        action:
          type: string
          enum: [approve, reject]
        note:
          type: string
          nullable: true
      required: [action]

    SiteSettings:
      type: object
      properties:
        maintenanceMode:
          type: boolean
        maintenanceMessage:
          type: string
        tasksBlocked:
          type: boolean
        withdrawalsBlocked:
          type: boolean
        depositAddress:
          type: string
      required: [maintenanceMode, maintenanceMessage, tasksBlocked, withdrawalsBlocked, depositAddress]