# lib/api-zod/src/generated/types/addBonusRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskCoin API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddBonusRequest {
  amount: number;
  note?: string;
}